chore: move keyManagement.js to wallet/ui/src
