+++ packages/wallet/ui/src/gen/google/protobuf/descriptor.ts
