+++ packages/wallet/ui/src/util/keyManagement.js
@@ -15,10 +15,10 @@ import { GenericAuthorization } from 'cosmjs-types/cosmos/authz/v1beta1/authz';
 import { QueryClientImpl } from 'cosmjs-types/cosmos/authz/v1beta1/query';
 
 // generated via ts-proto, as is conventional for cosmjs
-import { MsgWalletAction } from './gen/swingset/msgs';
+import { MsgWalletAction } from '../gen/swingset/msgs.js';
 
 // TODO: export these from '../../src/util/SuggestChain.js'
-import { stableCurrency, bech32Config } from './chainInfo.js';
+import { stableCurrency, bech32Config } from './SuggestChain.js';
 
 const { freeze } = Object;
